[#71] refactor: 블록 모달창 타입 리스트 항목 컴포넌트화.
BlockModalContentType.tsx 파일 추가
